requirements.txt
Django==4.1.7
djangorestframework==3.12.4
drf-spectacular==0.26.2
django-import-export==3.2.0
fontawesomefree==6.4.0
pytz